+++ src/zepben/ewb/model/cim/extensions/iec61968/common/contact_method_type.py
@@ -12,6 +12,15 @@ from zepben.ewb.model.cim.extensions.zbex import zbex
 
 @zbex
 class ContactMethodType(Enum):
+    """
+    [ZBEX] The method to use to make contact with a person or company.
+
+    :var UNKNOWN: [ZBEX] Unknown contact method type.
+    :var EMAIL: [ZBEX] Contact via email using the primary email address.
+    :var CALL: [ZBEX] Contact by call using the primary phone number.
+    :var LETTER: [ZBEX] Letter by post to the contact address is the method of contact.
+    """
+
     UNKNOWN = 0
     EMAIL = 1
     CALL = 2